+++ packages/fabric/editor-common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/editor-common
 
+## 5.1.1
+- [patch] Add Serializer for Single image [03405bf](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/03405bf)
+- [patch] Add Serializer for Single image [03405bf](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/03405bf)
+
 ## 5.1.0
 
 - [minor] FS-1461 added ContextIdentifierProvider interface to editor [0aeea41](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0aeea41)